Add localization docs and things
Introduces a new directory `docs/localization` to guide translators and
developers working on multi-language support for Onetime Secret. This is
a centralized and visible place to reference other content in the repo
like the language-specific language guis and glossaries in the src
directory (e.g. src/content/docs/uk/translations).